trusty: Don't disable preemption in nop kthread
A lot of time is spent in here with preemption disabled, hurting system-
wide latency. It's not necessary since the nop kthreads are pinned to their
respective CPUs, so any per-CPU data accessed in a nop kthread is safe
without disabling preemption.

Don't disable preemption in the nop kthread to improve system-wide latency.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Dark-Matter7232 <me@const.eu.org>